refactor(svelte): split eight oversize lib modules under the 500-LOC ratchet
Pure mechanical splits, no behavior change; facades and factory signatures
unchanged (zero callsite edits):

- index.ts (1086 -> 476): comma-packed one-line export statements at oxfmt's
  100-char width; geom/scale child generators updated to emit the compact
  form; 765-export (name, kind, lifecycle) surface byte-identical
- plot-engine.svelte.ts (883 -> 494) + diagnostics/plot-engine-diagnostics
  (274) + plot-engine-deriveds (304); construction DAG preserved
- surface-state (717 -> 292) + surface-handlers (497, live ports)
- inspection-state (585 -> 499) + inspection-reconcile (86) + traversal (76)
- interval-state (583 -> 453) + interval-bounds-state (180) + helpers (40)
- focus-state (543 -> 459) + focus-reconcile (177)
- GGPlot.svelte (570 -> 487) + tooltip-presence.svelte.ts (161)
- scene/Batch.svelte (526 -> 252) + batch-presentation.svelte.ts (366)

Rune rules honored (sync factories, no $state destructuring, live getters,
init-time effects, no module-level $state); autofixer clean on edited
modules. Ratchet pruned 51 -> 43. Gates: svelte-check 0/0, targeted browser
suites green (273/954/366/570/1008/513), test:ssr 132/132, lifecycle:check,
bun run check, knip, lint:type-aware:run.
