+++ packages/svelte/src/lib/index.ts
@@ -10,179 +10,91 @@
 /* oxlint-disable typescript/no-deprecated -- migration aliases intentionally remain public */
 
 // Components
-/** @lifecycle stable-intent */
-export { default as GGPlot } from "./GGPlot.svelte";
+/** @lifecycle stable-intent */ export { default as GGPlot } from "./GGPlot.svelte";
 export { default as Tooltip } from "./inspection/Tooltip.svelte";
-/** @lifecycle experimental */
 export { default as Inspect } from "./inspection/Inspect.svelte";
-export type {
-  LegendFilterClause,
-  LegendFilterEvent,
-  LegendFilterInput,
-  LegendFilterOptions,
-} from "./legend/filter.js";
+export type { LegendFilterClause, LegendFilterOptions } from "./legend/filter.js";
+export type { LegendFilterEvent, LegendFilterInput } from "./legend/filter.js";
 export { createPlotInteraction } from "./interaction/controller.svelte.js";
-export type {
-  ControllerDatumIdentity,
-  CreatePlotInteractionOptions,
-  PlotInteractionController,
-  PlotInteractionMutationOptions,
-  PlotInteractionZoomOptions,
-} from "./interaction/controller.svelte.js";
-export {
-  INTERACTION_DIAGNOSTIC_CATALOG,
-  normalizeInteractionConfig,
-} from "./interaction/interaction.js";
-export type {
-  AreaMode,
-  BrushSelection,
-  FacetIntervalPreset,
-  InspectInput,
-  InspectMode,
-  InspectOptions,
-  InteractionDiagnostic,
-  InteractionDiagnosticCode,
-  InteractionSource,
-  InteractionTool,
-  IntervalSelection,
-  LegendFocusChange,
-  LegendFocusClear,
-  LegendFocusEvent,
-  LegendFocusInput,
-  LegendFocusOptions,
-  NonEmptyReadonlyArray,
-  PlotDatum,
-  PlotInspection,
-  PlotInspectionChange,
-  PlotInspectionClear,
-  PlotInteractionEvent,
-  PlotInteractionChange,
-  PlotInteractionInterval,
-  PlotInteractionScope,
-  PlotInteractionSnapshot,
-  PlotInteractionTransition,
-  PlotSelection,
-  PointSelection,
-  ReadonlyZoomDomains,
-  ReadonlyIntervalDomains,
-  ResolvedInspectMode,
-  ResolvedInteractionConfig,
-  SelectInput,
-  SelectOptions,
-  ScopedInteractionDomain,
-  ScopedInteractionInterval,
-  ScopedInteractionKeys,
-  SemanticIntervalAxis,
-  TooltipContext,
-  TooltipField,
-  ZoomDomains,
-  ZoomEvent,
-  ZoomInput,
-  ZoomOptions,
-} from "./interaction/interaction.js";
+export type { PlotInteractionMutationOptions } from "./interaction/controller.svelte.js";
+export type { CreatePlotInteractionOptions } from "./interaction/controller.svelte.js";
+export type { PlotInteractionZoomOptions } from "./interaction/controller.svelte.js";
+export type { PlotInteractionController } from "./interaction/controller.svelte.js";
+export type { ControllerDatumIdentity } from "./interaction/controller.svelte.js";
+export { INTERACTION_DIAGNOSTIC_CATALOG } from "./interaction/interaction.js";
+export { normalizeInteractionConfig } from "./interaction/interaction.js";
+export type { InteractionDiagnosticCode, PlotInspectionChange } from "./interaction/interaction.js";
+export type { PlotInteractionEvent, PlotInteractionTransition } from "./interaction/interaction.js";
+export type { PlotInteractionScope, ResolvedInteractionConfig } from "./interaction/interaction.js";
+export type { ScopedInteractionInterval, SemanticIntervalAxis } from "./interaction/interaction.js";
+export type { InteractionDiagnostic, PlotInteractionInterval } from "./interaction/interaction.js";
+export type { NonEmptyReadonlyArray, PlotInteractionSnapshot } from "./interaction/interaction.js";
+export type { PlotInteractionChange, ReadonlyIntervalDomains } from "./interaction/interaction.js";
+export type { ScopedInteractionDomain, ScopedInteractionKeys } from "./interaction/interaction.js";
+export type { FacetIntervalPreset, PlotInspectionClear } from "./interaction/interaction.js";
+export type { ReadonlyZoomDomains, ResolvedInspectMode } from "./interaction/interaction.js";
+export type { AreaMode, InteractionSource, LegendFocusOptions } from "./interaction/interaction.js";
+export type { IntervalSelection, LegendFocusChange, PlotDatum } from "./interaction/interaction.js";
+export type { InspectMode, LegendFocusClear, LegendFocusEvent } from "./interaction/interaction.js";
+export type { InspectInput, InteractionTool, LegendFocusInput } from "./interaction/interaction.js";
+export type { BrushSelection, InspectOptions, PlotInspection } from "./interaction/interaction.js";
+export type { PlotSelection, PointSelection, TooltipContext } from "./interaction/interaction.js";
+export type { SelectInput, SelectOptions, TooltipField } from "./interaction/interaction.js";
+export type { ZoomDomains, ZoomEvent, ZoomInput, ZoomOptions } from "./interaction/interaction.js";
 // Geom* declaration-only children (#705 / #1039) — stable-intent: the
 // recommended Svelte composition surface for layers. Generated by
 // scripts/gen-geom-children.ts from KNOWN_GEOMS + *LayerInput types.
 // Registry/factory helpers stay experimental (file default).
 // <generated:geom-children> — bun run geom:children:gen
-/** @lifecycle stable-intent */
-export { default as GeomPoint } from "./geoms/GeomPoint.svelte";
-/** @lifecycle stable-intent */
-export { default as GeomCount } from "./geoms/GeomCount.svelte";
-/** @lifecycle stable-intent */
-export { default as GeomContour } from "./geoms/GeomContour.svelte";
-/** @lifecycle stable-intent */
-export { default as GeomDotplot } from "./geoms/GeomDotplot.svelte";
-/** @lifecycle stable-intent */
-export { default as GeomLine } from "./geoms/GeomLine.svelte";
-/** @lifecycle stable-intent */
-export { default as GeomPath } from "./geoms/GeomPath.svelte";
-/** @lifecycle stable-intent */
-export { default as GeomCol } from "./geoms/GeomCol.svelte";
-/** @lifecycle stable-intent */
-export { default as GeomBar } from "./geoms/GeomBar.svelte";
-/** @lifecycle stable-intent */
-export { default as GeomArea } from "./geoms/GeomArea.svelte";
-/** @lifecycle stable-intent */
-export { default as GeomRule } from "./geoms/GeomRule.svelte";
-/** @lifecycle stable-intent */
-export { default as GeomHline } from "./geoms/GeomHline.svelte";
-/** @lifecycle stable-intent */
-export { default as GeomVline } from "./geoms/GeomVline.svelte";
-/** @lifecycle stable-intent */
-export { default as GeomJitter } from "./geoms/GeomJitter.svelte";
-/** @lifecycle stable-intent */
-export { default as GeomText } from "./geoms/GeomText.svelte";
-/** @lifecycle stable-intent */
-export { default as GeomLabel } from "./geoms/GeomLabel.svelte";
-/** @lifecycle stable-intent */
-export { default as GeomHistogram } from "./geoms/GeomHistogram.svelte";
-/** @lifecycle stable-intent */
-export { default as GeomFreqpoly } from "./geoms/GeomFreqpoly.svelte";
-/** @lifecycle stable-intent */
-export { default as GeomSmooth } from "./geoms/GeomSmooth.svelte";
-/** @lifecycle stable-intent */
-export { default as GeomQuantile } from "./geoms/GeomQuantile.svelte";
-/** @lifecycle stable-intent */
-export { default as GeomBoxplot } from "./geoms/GeomBoxplot.svelte";
-/** @lifecycle stable-intent */
-export { default as GeomDensity } from "./geoms/GeomDensity.svelte";
-/** @lifecycle stable-intent */
-export { default as GeomDensity2d } from "./geoms/GeomDensity2d.svelte";
-/** @lifecycle stable-intent */
-export { default as GeomDensity2dFilled } from "./geoms/GeomDensity2dFilled.svelte";
-/** @lifecycle stable-intent */
-export { default as GeomErrorbar } from "./geoms/GeomErrorbar.svelte";
-/** @lifecycle stable-intent */
-export { default as GeomLinerange } from "./geoms/GeomLinerange.svelte";
-/** @lifecycle stable-intent */
-export { default as GeomPointrange } from "./geoms/GeomPointrange.svelte";
-/** @lifecycle stable-intent */
-export { default as GeomCrossbar } from "./geoms/GeomCrossbar.svelte";
-/** @lifecycle stable-intent */
-export { default as GeomRect } from "./geoms/GeomRect.svelte";
-/** @lifecycle stable-intent */
-export { default as GeomTile } from "./geoms/GeomTile.svelte";
-/** @lifecycle stable-intent */
-export { default as GeomBin2d } from "./geoms/GeomBin2d.svelte";
-/** @lifecycle stable-intent */
-export { default as GeomRaster } from "./geoms/GeomRaster.svelte";
-/** @lifecycle stable-intent */
-export { default as GeomHex } from "./geoms/GeomHex.svelte";
-/** @lifecycle stable-intent */
-export { default as GeomRibbon } from "./geoms/GeomRibbon.svelte";
-/** @lifecycle stable-intent */
-export { default as GeomSegment } from "./geoms/GeomSegment.svelte";
-/** @lifecycle stable-intent */
-export { default as GeomViolin } from "./geoms/GeomViolin.svelte";
-/** @lifecycle stable-intent */
-export { default as GeomFunction } from "./geoms/GeomFunction.svelte";
-/** @lifecycle stable-intent */
-export { default as GeomPolygon } from "./geoms/GeomPolygon.svelte";
-/** @lifecycle stable-intent */
-export { default as GeomAbline } from "./geoms/GeomAbline.svelte";
-/** @lifecycle stable-intent */
-export { default as GeomCurve } from "./geoms/GeomCurve.svelte";
-/** @lifecycle stable-intent */
-export { default as GeomMap } from "./geoms/GeomMap.svelte";
-/** @lifecycle stable-intent */
-export { default as GeomSf } from "./geoms/GeomSf.svelte";
-/** @lifecycle stable-intent */
-export { default as GeomSfText } from "./geoms/GeomSfText.svelte";
-/** @lifecycle stable-intent */
-export { default as GeomSfLabel } from "./geoms/GeomSfLabel.svelte";
-/** @lifecycle stable-intent */
-export { default as GeomBlank } from "./geoms/GeomBlank.svelte";
-/** @lifecycle stable-intent */
-export { default as GeomSpoke } from "./geoms/GeomSpoke.svelte";
-/** @lifecycle stable-intent */
-export { default as GeomRug } from "./geoms/GeomRug.svelte";
-/** @lifecycle stable-intent */
-export { default as GeomStep } from "./geoms/GeomStep.svelte";
-/** @lifecycle stable-intent */
-export { default as GeomQq } from "./geoms/GeomQq.svelte";
-/** @lifecycle stable-intent */
-export { default as GeomQqLine } from "./geoms/GeomQqLine.svelte";
+/** @lifecycle stable-intent */ export { default as GeomPoint } from "./geoms/GeomPoint.svelte";
+/** @lifecycle stable-intent */ export { default as GeomCount } from "./geoms/GeomCount.svelte";
+/** @lifecycle stable-intent */ export { default as GeomContour } from "./geoms/GeomContour.svelte";
+/** @lifecycle stable-intent */ export { default as GeomDotplot } from "./geoms/GeomDotplot.svelte";
+/** @lifecycle stable-intent */ export { default as GeomLine } from "./geoms/GeomLine.svelte";
+/** @lifecycle stable-intent */ export { default as GeomPath } from "./geoms/GeomPath.svelte";
+/** @lifecycle stable-intent */ export { default as GeomCol } from "./geoms/GeomCol.svelte";
+/** @lifecycle stable-intent */ export { default as GeomBar } from "./geoms/GeomBar.svelte";
+/** @lifecycle stable-intent */ export { default as GeomArea } from "./geoms/GeomArea.svelte";
+/** @lifecycle stable-intent */ export { default as GeomRule } from "./geoms/GeomRule.svelte";
+/** @lifecycle stable-intent */ export { default as GeomHline } from "./geoms/GeomHline.svelte";
+/** @lifecycle stable-intent */ export { default as GeomVline } from "./geoms/GeomVline.svelte";
+/** @lifecycle stable-intent */ export { default as GeomJitter } from "./geoms/GeomJitter.svelte";
+/** @lifecycle stable-intent */ export { default as GeomText } from "./geoms/GeomText.svelte";
+/** @lifecycle stable-intent */ export { default as GeomLabel } from "./geoms/GeomLabel.svelte";
+/** @lifecycle stable-intent */ export { default as GeomHistogram } from "./geoms/GeomHistogram.svelte";
+/** @lifecycle stable-intent */ export { default as GeomFreqpoly } from "./geoms/GeomFreqpoly.svelte";
+/** @lifecycle stable-intent */ export { default as GeomSmooth } from "./geoms/GeomSmooth.svelte";
+/** @lifecycle stable-intent */ export { default as GeomQuantile } from "./geoms/GeomQuantile.svelte";
+/** @lifecycle stable-intent */ export { default as GeomBoxplot } from "./geoms/GeomBoxplot.svelte";
+/** @lifecycle stable-intent */ export { default as GeomDensity } from "./geoms/GeomDensity.svelte";
+/** @lifecycle stable-intent */ export { default as GeomDensity2d } from "./geoms/GeomDensity2d.svelte";
+/** @lifecycle stable-intent */ export { default as GeomDensity2dFilled } from "./geoms/GeomDensity2dFilled.svelte";
+/** @lifecycle stable-intent */ export { default as GeomErrorbar } from "./geoms/GeomErrorbar.svelte";
+/** @lifecycle stable-intent */ export { default as GeomLinerange } from "./geoms/GeomLinerange.svelte";
+/** @lifecycle stable-intent */ export { default as GeomPointrange } from "./geoms/GeomPointrange.svelte";
+/** @lifecycle stable-intent */ export { default as GeomCrossbar } from "./geoms/GeomCrossbar.svelte";
+/** @lifecycle stable-intent */ export { default as GeomRect } from "./geoms/GeomRect.svelte";
+/** @lifecycle stable-intent */ export { default as GeomTile } from "./geoms/GeomTile.svelte";
+/** @lifecycle stable-intent */ export { default as GeomBin2d } from "./geoms/GeomBin2d.svelte";
+/** @lifecycle stable-intent */ export { default as GeomRaster } from "./geoms/GeomRaster.svelte";
+/** @lifecycle stable-intent */ export { default as GeomHex } from "./geoms/GeomHex.svelte";
+/** @lifecycle stable-intent */ export { default as GeomRibbon } from "./geoms/GeomRibbon.svelte";
+/** @lifecycle stable-intent */ export { default as GeomSegment } from "./geoms/GeomSegment.svelte";
+/** @lifecycle stable-intent */ export { default as GeomViolin } from "./geoms/GeomViolin.svelte";
+/** @lifecycle stable-intent */ export { default as GeomFunction } from "./geoms/GeomFunction.svelte";
+/** @lifecycle stable-intent */ export { default as GeomPolygon } from "./geoms/GeomPolygon.svelte";
+/** @lifecycle stable-intent */ export { default as GeomAbline } from "./geoms/GeomAbline.svelte";
+/** @lifecycle stable-intent */ export { default as GeomCurve } from "./geoms/GeomCurve.svelte";
+/** @lifecycle stable-intent */ export { default as GeomMap } from "./geoms/GeomMap.svelte";
+/** @lifecycle stable-intent */ export { default as GeomSf } from "./geoms/GeomSf.svelte";
+/** @lifecycle stable-intent */ export { default as GeomSfText } from "./geoms/GeomSfText.svelte";
+/** @lifecycle stable-intent */ export { default as GeomSfLabel } from "./geoms/GeomSfLabel.svelte";
+/** @lifecycle stable-intent */ export { default as GeomBlank } from "./geoms/GeomBlank.svelte";
+/** @lifecycle stable-intent */ export { default as GeomSpoke } from "./geoms/GeomSpoke.svelte";
+/** @lifecycle stable-intent */ export { default as GeomRug } from "./geoms/GeomRug.svelte";
+/** @lifecycle stable-intent */ export { default as GeomStep } from "./geoms/GeomStep.svelte";
+/** @lifecycle stable-intent */ export { default as GeomQq } from "./geoms/GeomQq.svelte";
+/** @lifecycle stable-intent */ export { default as GeomQqLine } from "./geoms/GeomQqLine.svelte";
 // </generated:geom-children>
 export { registerLayer, registerPlotLayer } from "./geoms/registry.svelte.js";
 export type { Layer, LayerRegistry, MarkLayerDescriptor } from "./geoms/registry.svelte.js";
@@ -191,375 +103,199 @@ export type { GeomProps } from "./geoms/factory.svelte.js";
 
 // Theme children (#659 slice 2) — stable-intent: 1:1 mirror of ThemeSpec /
 // THEME_NAMES. The GGPlot `theme` prop was removed in 0.13.0 (#704).
-/** @lifecycle stable-intent */
-export { default as Theme } from "./theme/Theme.svelte";
-/** @lifecycle stable-intent */
-export { default as ThemeDefault } from "./theme/ThemeDefault.svelte";
-/** @lifecycle stable-intent */
-export { default as ThemeLight } from "./theme/ThemeLight.svelte";
-/** @lifecycle stable-intent */
-export { default as ThemeDark } from "./theme/ThemeDark.svelte";
-/** @lifecycle stable-intent */
-export { default as ThemeMinimal } from "./theme/ThemeMinimal.svelte";
-/** @lifecycle stable-intent */
-export { default as ThemeGgplot2 } from "./theme/ThemeGgplot2.svelte";
-/** @lifecycle stable-intent */
-export { default as ThemeClassic } from "./theme/ThemeClassic.svelte";
-/** @lifecycle stable-intent */
-export { default as ThemeBw } from "./theme/ThemeBw.svelte";
-/** @lifecycle stable-intent */
-export { default as ThemeHrbr } from "./theme/ThemeHrbr.svelte";
-/** @lifecycle stable-intent */
-export { default as ThemeFew } from "./theme/ThemeFew.svelte";
-/** @lifecycle stable-intent */
-export { default as ThemeClean } from "./theme/ThemeClean.svelte";
-/** @lifecycle stable-intent */
-export { default as ThemeFivethirtyeight } from "./theme/ThemeFivethirtyeight.svelte";
-/** @lifecycle stable-intent */
-export { default as ThemeEconomist } from "./theme/ThemeEconomist.svelte";
-/** @lifecycle stable-intent */
-export { default as ThemeTufte } from "./theme/ThemeTufte.svelte";
-/** @lifecycle stable-intent */
-export { default as ThemeLinedraw } from "./theme/ThemeLinedraw.svelte";
-/** @lifecycle stable-intent */
-export { default as ThemeVoid } from "./theme/ThemeVoid.svelte";
-/** @lifecycle stable-intent */
-export { default as ThemeStata } from "./theme/ThemeStata.svelte";
-/** @lifecycle stable-intent */
-export { default as ThemeStatas1color } from "./theme/ThemeStatas1color.svelte";
-/** @lifecycle stable-intent */
-export { default as ThemeSolarized } from "./theme/ThemeSolarized.svelte";
-/** @lifecycle stable-intent */
-export { default as ThemeSolarizeddark } from "./theme/ThemeSolarizeddark.svelte";
-/** @lifecycle stable-intent */
-export { default as ThemeEconomistwhite } from "./theme/ThemeEconomistwhite.svelte";
-/** @lifecycle stable-intent */
-export { default as ThemeSolarized2 } from "./theme/ThemeSolarized2.svelte";
-/** @lifecycle stable-intent */
-export { default as ThemeSolarized2dark } from "./theme/ThemeSolarized2dark.svelte";
-/** @lifecycle stable-intent */
-export { default as ThemeWsj } from "./theme/ThemeWsj.svelte";
-/** @lifecycle stable-intent */
-/** @lifecycle stable-intent */
-export { default as ThemeHc } from "./theme/ThemeHc.svelte";
-/** @lifecycle stable-intent */
-export { default as ThemeHcdark } from "./theme/ThemeHcdark.svelte";
-/** @lifecycle stable-intent */
-export { default as ThemePander } from "./theme/ThemePander.svelte";
-/** @lifecycle stable-intent */
-export { default as ThemeBase } from "./theme/ThemeBase.svelte";
-/** @lifecycle stable-intent */
-export { default as ThemeIgray } from "./theme/ThemeIgray.svelte";
-/** @lifecycle stable-intent */
-export { default as ThemeMap } from "./theme/ThemeMap.svelte";
-/** @lifecycle stable-intent */
-export { default as ThemeSolid } from "./theme/ThemeSolid.svelte";
-/** @lifecycle stable-intent */
-export { default as ThemeGrey } from "./theme/ThemeGrey.svelte";
-/** @lifecycle stable-intent */
-export { default as ThemeGray } from "./theme/ThemeGray.svelte";
-/** @lifecycle stable-intent */
-export { default as ThemeTest } from "./theme/ThemeTest.svelte";
+/** @lifecycle stable-intent */ export { default as Theme } from "./theme/Theme.svelte";
+/** @lifecycle stable-intent */ export { default as ThemeDefault } from "./theme/ThemeDefault.svelte";
+/** @lifecycle stable-intent */ export { default as ThemeLight } from "./theme/ThemeLight.svelte";
+/** @lifecycle stable-intent */ export { default as ThemeDark } from "./theme/ThemeDark.svelte";
+/** @lifecycle stable-intent */ export { default as ThemeMinimal } from "./theme/ThemeMinimal.svelte";
+/** @lifecycle stable-intent */ export { default as ThemeGgplot2 } from "./theme/ThemeGgplot2.svelte";
+/** @lifecycle stable-intent */ export { default as ThemeClassic } from "./theme/ThemeClassic.svelte";
+/** @lifecycle stable-intent */ export { default as ThemeBw } from "./theme/ThemeBw.svelte";
+/** @lifecycle stable-intent */ export { default as ThemeHrbr } from "./theme/ThemeHrbr.svelte";
+/** @lifecycle stable-intent */ export { default as ThemeFew } from "./theme/ThemeFew.svelte";
+/** @lifecycle stable-intent */ export { default as ThemeClean } from "./theme/ThemeClean.svelte";
+/** @lifecycle stable-intent */ export { default as ThemeFivethirtyeight } from "./theme/ThemeFivethirtyeight.svelte";
+/** @lifecycle stable-intent */ export { default as ThemeEconomist } from "./theme/ThemeEconomist.svelte";
+/** @lifecycle stable-intent */ export { default as ThemeTufte } from "./theme/ThemeTufte.svelte";
+/** @lifecycle stable-intent */ export { default as ThemeLinedraw } from "./theme/ThemeLinedraw.svelte";
+/** @lifecycle stable-intent */ export { default as ThemeVoid } from "./theme/ThemeVoid.svelte";
+/** @lifecycle stable-intent */ export { default as ThemeStata } from "./theme/ThemeStata.svelte";
+/** @lifecycle stable-intent */ export { default as ThemeStatas1color } from "./theme/ThemeStatas1color.svelte";
+/** @lifecycle stable-intent */ export { default as ThemeSolarized } from "./theme/ThemeSolarized.svelte";
+/** @lifecycle stable-intent */ export { default as ThemeSolarizeddark } from "./theme/ThemeSolarizeddark.svelte";
+/** @lifecycle stable-intent */ export { default as ThemeEconomistwhite } from "./theme/ThemeEconomistwhite.svelte";
+/** @lifecycle stable-intent */ export { default as ThemeSolarized2 } from "./theme/ThemeSolarized2.svelte";
+/** @lifecycle stable-intent */ export { default as ThemeSolarized2dark } from "./theme/ThemeSolarized2dark.svelte";
+/** @lifecycle stable-intent */ export { default as ThemeWsj } from "./theme/ThemeWsj.svelte";
+/** @lifecycle stable-intent */ export { default as ThemeHc } from "./theme/ThemeHc.svelte";
+/** @lifecycle stable-intent */ export { default as ThemeHcdark } from "./theme/ThemeHcdark.svelte";
+/** @lifecycle stable-intent */ export { default as ThemePander } from "./theme/ThemePander.svelte";
+/** @lifecycle stable-intent */ export { default as ThemeBase } from "./theme/ThemeBase.svelte";
+/** @lifecycle stable-intent */ export { default as ThemeIgray } from "./theme/ThemeIgray.svelte";
+/** @lifecycle stable-intent */ export { default as ThemeMap } from "./theme/ThemeMap.svelte";
+/** @lifecycle stable-intent */ export { default as ThemeSolid } from "./theme/ThemeSolid.svelte";
+/** @lifecycle stable-intent */ export { default as ThemeGrey } from "./theme/ThemeGrey.svelte";
+/** @lifecycle stable-intent */ export { default as ThemeGray } from "./theme/ThemeGray.svelte";
+/** @lifecycle stable-intent */ export { default as ThemeTest } from "./theme/ThemeTest.svelte";
 
 // Scale children (#659) — stable-intent shells for every SCALE_CAPABILITIES
 // family + <Scale> escape hatch; destination for the deprecated `scales` prop
 // (since 0.11.0). Generated shells live between the markers; run
 // `bun run scale:children:gen` to regenerate. Colour aliases re-export Color.
-/** @lifecycle stable-intent */
-export { default as Scale } from "./scale/Scale.svelte";
+/** @lifecycle stable-intent */ export { default as Scale } from "./scale/Scale.svelte";
 // <generated:scale-children> — bun run scale:children:gen
-/** @lifecycle stable-intent */
-export { default as ScaleXContinuous } from "./scale/ScaleXContinuous.svelte";
-/** @lifecycle stable-intent */
-export { default as ScaleYContinuous } from "./scale/ScaleYContinuous.svelte";
-/** @lifecycle stable-intent */
-export { default as ScaleXLog10 } from "./scale/ScaleXLog10.svelte";
-/** @lifecycle stable-intent */
-export { default as ScaleYLog10 } from "./scale/ScaleYLog10.svelte";
-/** @lifecycle stable-intent */
-export { default as ScaleXSqrt } from "./scale/ScaleXSqrt.svelte";
-/** @lifecycle stable-intent */
-export { default as ScaleYSqrt } from "./scale/ScaleYSqrt.svelte";
-/** @lifecycle stable-intent */
-export { default as ScaleXReverse } from "./scale/ScaleXReverse.svelte";
-/** @lifecycle stable-intent */
-export { default as ScaleYReverse } from "./scale/ScaleYReverse.svelte";
-/** @lifecycle stable-intent */
-export { default as ScaleXBinned } from "./scale/ScaleXBinned.svelte";
-/** @lifecycle stable-intent */
-export { default as ScaleYBinned } from "./scale/ScaleYBinned.svelte";
-/** @lifecycle stable-intent */
-export { default as ScaleXDate } from "./scale/ScaleXDate.svelte";
-/** @lifecycle stable-intent */
-export { default as ScaleXDatetime } from "./scale/ScaleXDatetime.svelte";
-/** @lifecycle stable-intent */
-export { default as ScaleXTime } from "./scale/ScaleXTime.svelte";
-/** @lifecycle stable-intent */
-export { default as ScaleYDate } from "./scale/ScaleYDate.svelte";
-/** @lifecycle stable-intent */
-export { default as ScaleYDatetime } from "./scale/ScaleYDatetime.svelte";
-/** @lifecycle stable-intent */
-export { default as ScaleYTime } from "./scale/ScaleYTime.svelte";
-/** @lifecycle stable-intent */
-export { default as ScaleXMonthDay } from "./scale/ScaleXMonthDay.svelte";
-/** @lifecycle stable-intent */
-export { default as ScaleYMonthDay } from "./scale/ScaleYMonthDay.svelte";
-/** @lifecycle stable-intent */
-export { default as ScaleXDiscrete } from "./scale/ScaleXDiscrete.svelte";
-/** @lifecycle stable-intent */
-export { default as ScaleYDiscrete } from "./scale/ScaleYDiscrete.svelte";
-/** @lifecycle stable-intent */
-export { default as ScaleColorContinuous } from "./scale/ScaleColorContinuous.svelte";
-/** @lifecycle stable-intent */
-export { default as ScaleColorDiscrete } from "./scale/ScaleColorDiscrete.svelte";
-/** @lifecycle stable-intent */
-export { default as ScaleColorBinned } from "./scale/ScaleColorBinned.svelte";
-/** @lifecycle stable-intent */
-export { default as ScaleColorBrewer } from "./scale/ScaleColorBrewer.svelte";
-/** @lifecycle stable-intent */
-export { default as ScaleColorDistiller } from "./scale/ScaleColorDistiller.svelte";
-/** @lifecycle stable-intent */
-export { default as ScaleColorFermenter } from "./scale/ScaleColorFermenter.svelte";
-/** @lifecycle stable-intent */
-export { default as ScaleColorSteps } from "./scale/ScaleColorSteps.svelte";
-/** @lifecycle stable-intent */
-export { default as ScaleColorSteps2 } from "./scale/ScaleColorSteps2.svelte";
-/** @lifecycle stable-intent */
-export { default as ScaleColorStepsn } from "./scale/ScaleColorStepsn.svelte";
-/** @lifecycle stable-intent */
-export { default as ScaleColorGradient } from "./scale/ScaleColorGradient.svelte";
-/** @lifecycle stable-intent */
-export { default as ScaleColorGradient2 } from "./scale/ScaleColorGradient2.svelte";
-/** @lifecycle stable-intent */
-export { default as ScaleColorGradientn } from "./scale/ScaleColorGradientn.svelte";
-/** @lifecycle stable-intent */
-export { default as ScaleColorHue } from "./scale/ScaleColorHue.svelte";
-/** @lifecycle stable-intent */
-export { default as ScaleColorGrey } from "./scale/ScaleColorGrey.svelte";
-/** @lifecycle stable-intent */
-export { default as ScaleColorOrdinal } from "./scale/ScaleColorOrdinal.svelte";
-/** @lifecycle stable-intent */
-export { default as ScaleColorLog10 } from "./scale/ScaleColorLog10.svelte";
-/** @lifecycle stable-intent */
-export { default as ScaleColorSqrt } from "./scale/ScaleColorSqrt.svelte";
-/** @lifecycle stable-intent */
-export { default as ScaleColorDate } from "./scale/ScaleColorDate.svelte";
-/** @lifecycle stable-intent */
-export { default as ScaleColorDatetime } from "./scale/ScaleColorDatetime.svelte";
-/** @lifecycle stable-intent */
-export { default as ScaleColorManual } from "./scale/ScaleColorManual.svelte";
-/** @lifecycle stable-intent */
-export { default as ScaleColorIdentity } from "./scale/ScaleColorIdentity.svelte";
-/** @lifecycle stable-intent */
-export { default as ScaleFillContinuous } from "./scale/ScaleFillContinuous.svelte";
-/** @lifecycle stable-intent */
-export { default as ScaleFillDiscrete } from "./scale/ScaleFillDiscrete.svelte";
-/** @lifecycle stable-intent */
-export { default as ScaleFillBinned } from "./scale/ScaleFillBinned.svelte";
-/** @lifecycle stable-intent */
-export { default as ScaleFillSteps } from "./scale/ScaleFillSteps.svelte";
-/** @lifecycle stable-intent */
-export { default as ScaleFillSteps2 } from "./scale/ScaleFillSteps2.svelte";
-/** @lifecycle stable-intent */
-export { default as ScaleFillStepsn } from "./scale/ScaleFillStepsn.svelte";
-/** @lifecycle stable-intent */
-export { default as ScaleFillGradient } from "./scale/ScaleFillGradient.svelte";
-/** @lifecycle stable-intent */
-export { default as ScaleFillGradient2 } from "./scale/ScaleFillGradient2.svelte";
-/** @lifecycle stable-intent */
-export { default as ScaleFillGradientn } from "./scale/ScaleFillGradientn.svelte";
-/** @lifecycle stable-intent */
-export { default as ScaleFillHue } from "./scale/ScaleFillHue.svelte";
-/** @lifecycle stable-intent */
-export { default as ScaleFillGrey } from "./scale/ScaleFillGrey.svelte";
-/** @lifecycle stable-intent */
-export { default as ScaleFillOrdinal } from "./scale/ScaleFillOrdinal.svelte";
-/** @lifecycle stable-intent */
-export { default as ScaleFillLog10 } from "./scale/ScaleFillLog10.svelte";
-/** @lifecycle stable-intent */
-export { default as ScaleFillSqrt } from "./scale/ScaleFillSqrt.svelte";
-/** @lifecycle stable-intent */
-export { default as ScaleFillDate } from "./scale/ScaleFillDate.svelte";
-/** @lifecycle stable-intent */
-export { default as ScaleFillDatetime } from "./scale/ScaleFillDatetime.svelte";
-/** @lifecycle stable-intent */
-export { default as ScaleFillManual } from "./scale/ScaleFillManual.svelte";
-/** @lifecycle stable-intent */
-export { default as ScaleFillIdentity } from "./scale/ScaleFillIdentity.svelte";
-/** @lifecycle stable-intent */
-export { default as ScaleFillBrewer } from "./scale/ScaleFillBrewer.svelte";
-/** @lifecycle stable-intent */
-export { default as ScaleFillDistiller } from "./scale/ScaleFillDistiller.svelte";
-/** @lifecycle stable-intent */
-export { default as ScaleFillFermenter } from "./scale/ScaleFillFermenter.svelte";
-/** @lifecycle stable-intent */
-export { default as ScaleColorViridisC } from "./scale/ScaleColorViridisC.svelte";
-/** @lifecycle stable-intent */
-export { default as ScaleColorViridisD } from "./scale/ScaleColorViridisD.svelte";
-/** @lifecycle stable-intent */
-export { default as ScaleColorViridisB } from "./scale/ScaleColorViridisB.svelte";
-/** @lifecycle stable-intent */
-export { default as ScaleFillViridisC } from "./scale/ScaleFillViridisC.svelte";
-/** @lifecycle stable-intent */
-export { default as ScaleFillViridisD } from "./scale/ScaleFillViridisD.svelte";
-/** @lifecycle stable-intent */
-export { default as ScaleFillViridisB } from "./scale/ScaleFillViridisB.svelte";
-/** @lifecycle stable-intent */
-export { default as ScaleSizeContinuous } from "./scale/ScaleSizeContinuous.svelte";
-/** @lifecycle stable-intent */
-export { default as ScaleSizeDiscrete } from "./scale/ScaleSizeDiscrete.svelte";
-/** @lifecycle stable-intent */
-export { default as ScaleSizeBinned } from "./scale/ScaleSizeBinned.svelte";
-/** @lifecycle stable-intent */
-export { default as ScaleSizeDate } from "./scale/ScaleSizeDate.svelte";
-/** @lifecycle stable-intent */
-export { default as ScaleSizeDatetime } from "./scale/ScaleSizeDatetime.svelte";
-/** @lifecycle stable-intent */
-export { default as ScaleSizeManual } from "./scale/ScaleSizeManual.svelte";
-/** @lifecycle stable-intent */
-export { default as ScaleSizeIdentity } from "./scale/ScaleSizeIdentity.svelte";
-/** @lifecycle stable-intent */
-export { default as ScaleLinewidthContinuous } from "./scale/ScaleLinewidthContinuous.svelte";
-/** @lifecycle stable-intent */
-export { default as ScaleLinewidthDiscrete } from "./scale/ScaleLinewidthDiscrete.svelte";
-/** @lifecycle stable-intent */
-export { default as ScaleLinewidthBinned } from "./scale/ScaleLinewidthBinned.svelte";
-/** @lifecycle stable-intent */
-export { default as ScaleLinewidthDate } from "./scale/ScaleLinewidthDate.svelte";
-/** @lifecycle stable-intent */
-export { default as ScaleLinewidthDatetime } from "./scale/ScaleLinewidthDatetime.svelte";
-/** @lifecycle stable-intent */
-export { default as ScaleLinewidthManual } from "./scale/ScaleLinewidthManual.svelte";
-/** @lifecycle stable-intent */
-export { default as ScaleLinewidthIdentity } from "./scale/ScaleLinewidthIdentity.svelte";
-/** @lifecycle stable-intent */
-export { default as ScaleAlphaContinuous } from "./scale/ScaleAlphaContinuous.svelte";
-/** @lifecycle stable-intent */
-export { default as ScaleAlphaDiscrete } from "./scale/ScaleAlphaDiscrete.svelte";
-/** @lifecycle stable-intent */
-export { default as ScaleAlphaBinned } from "./scale/ScaleAlphaBinned.svelte";
-/** @lifecycle stable-intent */
-export { default as ScaleAlphaDate } from "./scale/ScaleAlphaDate.svelte";
-/** @lifecycle stable-intent */
-export { default as ScaleAlphaDatetime } from "./scale/ScaleAlphaDatetime.svelte";
-/** @lifecycle stable-intent */
-export { default as ScaleAlphaManual } from "./scale/ScaleAlphaManual.svelte";
-/** @lifecycle stable-intent */
-export { default as ScaleAlphaIdentity } from "./scale/ScaleAlphaIdentity.svelte";
-/** @lifecycle stable-intent */
-export { default as ScaleSizeArea } from "./scale/ScaleSizeArea.svelte";
-/** @lifecycle stable-intent */
-export { default as ScaleSizeBinnedArea } from "./scale/ScaleSizeBinnedArea.svelte";
-/** @lifecycle stable-intent */
-export { default as ScaleRadius } from "./scale/ScaleRadius.svelte";
-/** @lifecycle stable-intent */
-export { default as ScaleShapeDiscrete } from "./scale/ScaleShapeDiscrete.svelte";
-/** @lifecycle stable-intent */
-export { default as ScaleShapeBinned } from "./scale/ScaleShapeBinned.svelte";
-/** @lifecycle stable-intent */
-export { default as ScaleShapeManual } from "./scale/ScaleShapeManual.svelte";
-/** @lifecycle stable-intent */
-export { default as ScaleShapeIdentity } from "./scale/ScaleShapeIdentity.svelte";
-/** @lifecycle stable-intent */
-export { default as ScaleLinetypeDiscrete } from "./scale/ScaleLinetypeDiscrete.svelte";
-/** @lifecycle stable-intent */
-export { default as ScaleLinetypeBinned } from "./scale/ScaleLinetypeBinned.svelte";
-/** @lifecycle stable-intent */
-export { default as ScaleLinetypeManual } from "./scale/ScaleLinetypeManual.svelte";
-/** @lifecycle stable-intent */
-export { default as ScaleLinetypeIdentity } from "./scale/ScaleLinetypeIdentity.svelte";
-/** @lifecycle stable-intent */
-export { default as ScaleColourContinuous } from "./scale/ScaleColorContinuous.svelte";
-/** @lifecycle stable-intent */
-export { default as ScaleColourDiscrete } from "./scale/ScaleColorDiscrete.svelte";
-/** @lifecycle stable-intent */
-export { default as ScaleColourBinned } from "./scale/ScaleColorBinned.svelte";
-/** @lifecycle stable-intent */
-export { default as ScaleColourBrewer } from "./scale/ScaleColorBrewer.svelte";
-/** @lifecycle stable-intent */
-export { default as ScaleColourDistiller } from "./scale/ScaleColorDistiller.svelte";
-/** @lifecycle stable-intent */
-export { default as ScaleColourFermenter } from "./scale/ScaleColorFermenter.svelte";
-/** @lifecycle stable-intent */
-export { default as ScaleColourSteps } from "./scale/ScaleColorSteps.svelte";
-/** @lifecycle stable-intent */
-export { default as ScaleColourSteps2 } from "./scale/ScaleColorSteps2.svelte";
-/** @lifecycle stable-intent */
-export { default as ScaleColourStepsn } from "./scale/ScaleColorStepsn.svelte";
-/** @lifecycle stable-intent */
-export { default as ScaleColourGradient } from "./scale/ScaleColorGradient.svelte";
-/** @lifecycle stable-intent */
-export { default as ScaleColourGradient2 } from "./scale/ScaleColorGradient2.svelte";
-/** @lifecycle stable-intent */
-export { default as ScaleColourGradientn } from "./scale/ScaleColorGradientn.svelte";
-/** @lifecycle stable-intent */
-export { default as ScaleColourHue } from "./scale/ScaleColorHue.svelte";
-/** @lifecycle stable-intent */
-export { default as ScaleColourGrey } from "./scale/ScaleColorGrey.svelte";
-/** @lifecycle stable-intent */
-export { default as ScaleColourOrdinal } from "./scale/ScaleColorOrdinal.svelte";
-/** @lifecycle stable-intent */
-export { default as ScaleColourLog10 } from "./scale/ScaleColorLog10.svelte";
-/** @lifecycle stable-intent */
-export { default as ScaleColourSqrt } from "./scale/ScaleColorSqrt.svelte";
-/** @lifecycle stable-intent */
-export { default as ScaleColourDate } from "./scale/ScaleColorDate.svelte";
-/** @lifecycle stable-intent */
-export { default as ScaleColourDatetime } from "./scale/ScaleColorDatetime.svelte";
-/** @lifecycle stable-intent */
-export { default as ScaleColourManual } from "./scale/ScaleColorManual.svelte";
-/** @lifecycle stable-intent */
-export { default as ScaleColourIdentity } from "./scale/ScaleColorIdentity.svelte";
-/** @lifecycle stable-intent */
-export { default as ScaleColourViridisC } from "./scale/ScaleColorViridisC.svelte";
-/** @lifecycle stable-intent */
-export { default as ScaleColourViridisD } from "./scale/ScaleColorViridisD.svelte";
-/** @lifecycle stable-intent */
-export { default as ScaleColourViridisB } from "./scale/ScaleColorViridisB.svelte";
-/** @lifecycle stable-intent */
-export { default as ScaleSizeOrdinal } from "./scale/ScaleSizeDiscrete.svelte";
-/** @lifecycle stable-intent */
-export { default as ScaleLinewidthOrdinal } from "./scale/ScaleLinewidthDiscrete.svelte";
-/** @lifecycle stable-intent */
-export { default as ScaleAlphaOrdinal } from "./scale/ScaleAlphaDiscrete.svelte";
-/** @lifecycle stable-intent */
-export { default as ScaleShapeOrdinal } from "./scale/ScaleShapeDiscrete.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleXContinuous } from "./scale/ScaleXContinuous.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleYContinuous } from "./scale/ScaleYContinuous.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleXLog10 } from "./scale/ScaleXLog10.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleYLog10 } from "./scale/ScaleYLog10.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleXSqrt } from "./scale/ScaleXSqrt.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleYSqrt } from "./scale/ScaleYSqrt.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleXReverse } from "./scale/ScaleXReverse.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleYReverse } from "./scale/ScaleYReverse.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleXBinned } from "./scale/ScaleXBinned.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleYBinned } from "./scale/ScaleYBinned.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleXDate } from "./scale/ScaleXDate.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleXDatetime } from "./scale/ScaleXDatetime.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleXTime } from "./scale/ScaleXTime.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleYDate } from "./scale/ScaleYDate.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleYDatetime } from "./scale/ScaleYDatetime.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleYTime } from "./scale/ScaleYTime.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleXMonthDay } from "./scale/ScaleXMonthDay.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleYMonthDay } from "./scale/ScaleYMonthDay.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleXDiscrete } from "./scale/ScaleXDiscrete.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleYDiscrete } from "./scale/ScaleYDiscrete.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleColorContinuous } from "./scale/ScaleColorContinuous.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleColorDiscrete } from "./scale/ScaleColorDiscrete.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleColorBinned } from "./scale/ScaleColorBinned.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleColorBrewer } from "./scale/ScaleColorBrewer.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleColorDistiller } from "./scale/ScaleColorDistiller.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleColorFermenter } from "./scale/ScaleColorFermenter.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleColorSteps } from "./scale/ScaleColorSteps.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleColorSteps2 } from "./scale/ScaleColorSteps2.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleColorStepsn } from "./scale/ScaleColorStepsn.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleColorGradient } from "./scale/ScaleColorGradient.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleColorGradient2 } from "./scale/ScaleColorGradient2.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleColorGradientn } from "./scale/ScaleColorGradientn.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleColorHue } from "./scale/ScaleColorHue.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleColorGrey } from "./scale/ScaleColorGrey.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleColorOrdinal } from "./scale/ScaleColorOrdinal.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleColorLog10 } from "./scale/ScaleColorLog10.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleColorSqrt } from "./scale/ScaleColorSqrt.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleColorDate } from "./scale/ScaleColorDate.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleColorDatetime } from "./scale/ScaleColorDatetime.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleColorManual } from "./scale/ScaleColorManual.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleColorIdentity } from "./scale/ScaleColorIdentity.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleFillContinuous } from "./scale/ScaleFillContinuous.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleFillDiscrete } from "./scale/ScaleFillDiscrete.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleFillBinned } from "./scale/ScaleFillBinned.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleFillSteps } from "./scale/ScaleFillSteps.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleFillSteps2 } from "./scale/ScaleFillSteps2.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleFillStepsn } from "./scale/ScaleFillStepsn.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleFillGradient } from "./scale/ScaleFillGradient.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleFillGradient2 } from "./scale/ScaleFillGradient2.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleFillGradientn } from "./scale/ScaleFillGradientn.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleFillHue } from "./scale/ScaleFillHue.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleFillGrey } from "./scale/ScaleFillGrey.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleFillOrdinal } from "./scale/ScaleFillOrdinal.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleFillLog10 } from "./scale/ScaleFillLog10.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleFillSqrt } from "./scale/ScaleFillSqrt.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleFillDate } from "./scale/ScaleFillDate.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleFillDatetime } from "./scale/ScaleFillDatetime.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleFillManual } from "./scale/ScaleFillManual.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleFillIdentity } from "./scale/ScaleFillIdentity.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleFillBrewer } from "./scale/ScaleFillBrewer.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleFillDistiller } from "./scale/ScaleFillDistiller.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleFillFermenter } from "./scale/ScaleFillFermenter.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleColorViridisC } from "./scale/ScaleColorViridisC.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleColorViridisD } from "./scale/ScaleColorViridisD.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleColorViridisB } from "./scale/ScaleColorViridisB.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleFillViridisC } from "./scale/ScaleFillViridisC.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleFillViridisD } from "./scale/ScaleFillViridisD.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleFillViridisB } from "./scale/ScaleFillViridisB.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleSizeContinuous } from "./scale/ScaleSizeContinuous.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleSizeDiscrete } from "./scale/ScaleSizeDiscrete.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleSizeBinned } from "./scale/ScaleSizeBinned.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleSizeDate } from "./scale/ScaleSizeDate.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleSizeDatetime } from "./scale/ScaleSizeDatetime.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleSizeManual } from "./scale/ScaleSizeManual.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleSizeIdentity } from "./scale/ScaleSizeIdentity.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleLinewidthContinuous } from "./scale/ScaleLinewidthContinuous.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleLinewidthDiscrete } from "./scale/ScaleLinewidthDiscrete.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleLinewidthBinned } from "./scale/ScaleLinewidthBinned.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleLinewidthDate } from "./scale/ScaleLinewidthDate.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleLinewidthDatetime } from "./scale/ScaleLinewidthDatetime.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleLinewidthManual } from "./scale/ScaleLinewidthManual.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleLinewidthIdentity } from "./scale/ScaleLinewidthIdentity.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleAlphaContinuous } from "./scale/ScaleAlphaContinuous.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleAlphaDiscrete } from "./scale/ScaleAlphaDiscrete.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleAlphaBinned } from "./scale/ScaleAlphaBinned.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleAlphaDate } from "./scale/ScaleAlphaDate.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleAlphaDatetime } from "./scale/ScaleAlphaDatetime.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleAlphaManual } from "./scale/ScaleAlphaManual.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleAlphaIdentity } from "./scale/ScaleAlphaIdentity.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleSizeArea } from "./scale/ScaleSizeArea.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleSizeBinnedArea } from "./scale/ScaleSizeBinnedArea.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleRadius } from "./scale/ScaleRadius.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleShapeDiscrete } from "./scale/ScaleShapeDiscrete.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleShapeBinned } from "./scale/ScaleShapeBinned.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleShapeManual } from "./scale/ScaleShapeManual.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleShapeIdentity } from "./scale/ScaleShapeIdentity.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleLinetypeDiscrete } from "./scale/ScaleLinetypeDiscrete.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleLinetypeBinned } from "./scale/ScaleLinetypeBinned.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleLinetypeManual } from "./scale/ScaleLinetypeManual.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleLinetypeIdentity } from "./scale/ScaleLinetypeIdentity.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleColourContinuous } from "./scale/ScaleColorContinuous.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleColourDiscrete } from "./scale/ScaleColorDiscrete.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleColourBinned } from "./scale/ScaleColorBinned.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleColourBrewer } from "./scale/ScaleColorBrewer.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleColourDistiller } from "./scale/ScaleColorDistiller.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleColourFermenter } from "./scale/ScaleColorFermenter.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleColourSteps } from "./scale/ScaleColorSteps.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleColourSteps2 } from "./scale/ScaleColorSteps2.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleColourStepsn } from "./scale/ScaleColorStepsn.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleColourGradient } from "./scale/ScaleColorGradient.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleColourGradient2 } from "./scale/ScaleColorGradient2.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleColourGradientn } from "./scale/ScaleColorGradientn.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleColourHue } from "./scale/ScaleColorHue.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleColourGrey } from "./scale/ScaleColorGrey.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleColourOrdinal } from "./scale/ScaleColorOrdinal.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleColourLog10 } from "./scale/ScaleColorLog10.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleColourSqrt } from "./scale/ScaleColorSqrt.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleColourDate } from "./scale/ScaleColorDate.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleColourDatetime } from "./scale/ScaleColorDatetime.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleColourManual } from "./scale/ScaleColorManual.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleColourIdentity } from "./scale/ScaleColorIdentity.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleColourViridisC } from "./scale/ScaleColorViridisC.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleColourViridisD } from "./scale/ScaleColorViridisD.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleColourViridisB } from "./scale/ScaleColorViridisB.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleSizeOrdinal } from "./scale/ScaleSizeDiscrete.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleLinewidthOrdinal } from "./scale/ScaleLinewidthDiscrete.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleAlphaOrdinal } from "./scale/ScaleAlphaDiscrete.svelte";
+/** @lifecycle stable-intent */ export { default as ScaleShapeOrdinal } from "./scale/ScaleShapeDiscrete.svelte";
 // </generated:scale-children>
 
 // Coord children (#659 slice 5) — stable-intent: named shells + <Coord>
 // escape hatch; destination for the deprecated `coord` prop (since 0.11.0).
-/** @lifecycle stable-intent */
-export { default as Coord } from "./coord/Coord.svelte";
-/** @lifecycle stable-intent */
-export { default as CoordFlip } from "./coord/CoordFlip.svelte";
-/** @lifecycle stable-intent */
-export { default as CoordCartesian } from "./coord/CoordCartesian.svelte";
-/** @lifecycle stable-intent */
-export { default as CoordTransform } from "./coord/CoordTransform.svelte";
-/** @lifecycle stable-intent */
-export { default as CoordFixed } from "./coord/CoordFixed.svelte";
+/** @lifecycle stable-intent */ export { default as Coord } from "./coord/Coord.svelte";
+/** @lifecycle stable-intent */ export { default as CoordFlip } from "./coord/CoordFlip.svelte";
+/** @lifecycle stable-intent */ export { default as CoordCartesian } from "./coord/CoordCartesian.svelte";
+/** @lifecycle stable-intent */ export { default as CoordTransform } from "./coord/CoordTransform.svelte";
+/** @lifecycle stable-intent */ export { default as CoordFixed } from "./coord/CoordFixed.svelte";
 export { default as CoordSf } from "./coord/CoordSf.svelte";
-/** @lifecycle stable-intent */
-export { default as CoordEqual } from "./coord/CoordFixed.svelte";
-/** @lifecycle experimental */
+/** @lifecycle stable-intent */ export { default as CoordEqual } from "./coord/CoordFixed.svelte";
 export { default as CoordRadial } from "./coord/CoordRadial.svelte";
-/** @lifecycle experimental */
 export { default as CoordPolar } from "./coord/CoordPolar.svelte";
 
 // Facet children (#659 slice 5) — stable-intent: complete <Facet> surface +
 // ggplot2-spelling shells; destination for the deprecated `facet` prop.
-/** @lifecycle stable-intent */
-export { default as Facet } from "./facet/Facet.svelte";
-/** @lifecycle stable-intent */
-export { default as FacetWrap } from "./facet/FacetWrap.svelte";
-/** @lifecycle stable-intent */
-export { default as FacetGrid } from "./facet/FacetGrid.svelte";
+/** @lifecycle stable-intent */ export { default as Facet } from "./facet/Facet.svelte";
+/** @lifecycle stable-intent */ export { default as FacetWrap } from "./facet/FacetWrap.svelte";
+/** @lifecycle stable-intent */ export { default as FacetGrid } from "./facet/FacetGrid.svelte";
 
 // Labs child (#659 slice 6) — stable-intent: the whole flat Labs surface as
 // named props; destination for the deprecated `labs` prop (since 0.11.0).
 // No <Labs value> escape hatch — `<Labs {...computed} />` already covers it.
-/** @lifecycle stable-intent */
-export { default as Labs } from "./labs/Labs.svelte";
+/** @lifecycle stable-intent */ export { default as Labs } from "./labs/Labs.svelte";
 // Name collision, resolved in favour of the component: every grammar child is
 // named for the PortableSpec field it fills (Theme, Scale, Coord, Facet,
 // Guides, Legend), and Labs is the only spec type without a Spec/Input suffix.
@@ -570,517 +306,171 @@ export type { Labs as LabsSpec } from "@ggsvelte/spec";
 // Guide children (#659 slice 6) — stable-intent: one shell per guide TYPE,
 // each keyed by a `channel` prop (the aesthetic is a key, not a helper), plus
 // the <Guides value> escape hatch. Destination for the deprecated `guides` prop.
-/** @lifecycle stable-intent */
-export { default as Guides } from "./guides/Guides.svelte";
-/** @lifecycle stable-intent */
-export { default as GuideAxis } from "./guides/GuideAxis.svelte";
-/** @lifecycle stable-intent */
-export { default as GuideLegend } from "./guides/GuideLegend.svelte";
-/** @lifecycle stable-intent */
-export { default as GuideColorbar } from "./guides/GuideColorbar.svelte";
-/** @lifecycle stable-intent */
-export { default as GuideColorsteps } from "./guides/GuideColorsteps.svelte";
-/** @lifecycle stable-intent */
-export { default as GuideNone } from "./guides/GuideNone.svelte";
-/** @lifecycle stable-intent */
-export type {
-  GuideChannel,
-  NonPositionGuideChannel,
-  PositionGuideChannel,
-} from "./guides/factory.svelte.js";
+/** @lifecycle stable-intent */ export { default as Guides } from "./guides/Guides.svelte";
+/** @lifecycle stable-intent */ export { default as GuideAxis } from "./guides/GuideAxis.svelte";
+/** @lifecycle stable-intent */ export { default as GuideLegend } from "./guides/GuideLegend.svelte";
+/** @lifecycle stable-intent */ export { default as GuideColorbar } from "./guides/GuideColorbar.svelte";
+/** @lifecycle stable-intent */ export { default as GuideColorsteps } from "./guides/GuideColorsteps.svelte";
+/** @lifecycle stable-intent */ export { default as GuideNone } from "./guides/GuideNone.svelte";
+/** @lifecycle stable-intent */ export type { NonPositionGuideChannel } from "./guides/factory.svelte.js";
+/** @lifecycle stable-intent */ export type { PositionGuideChannel } from "./guides/factory.svelte.js";
+/** @lifecycle stable-intent */ export type { GuideChannel } from "./guides/factory.svelte.js";
 
 // Legend child (#659 slice 6) — stable-intent: the plot-wide entry-SORT enum.
 // Not the per-aesthetic placement rank; that is <GuideLegend order={…}/>.
-/** @lifecycle stable-intent */
-export { default as Legend } from "./legend/Legend.svelte";
+/** @lifecycle stable-intent */ export { default as Legend } from "./legend/Legend.svelte";
 
 // Deprecation + composition diagnostics (sibling unions to InteractionDiagnostic).
 // Grammar-prop emission left in 0.13.0 (#704); catalog + builder remain for
 // PlotDiagnostic union typing, upgrade-guide anchors, and codemod consumers.
 // Discriminate with `d.code === "DEPRECATED_PLOT_PROP"` if needed.
-export {
-  DEPRECATION_DIAGNOSTIC_CATALOG,
-  deprecatedPropDiagnostic,
-} from "./diagnostics/deprecation.js";
-export type {
-  DeprecationDiagnostic,
-  DeprecationDiagnosticCode,
-  PlotDiagnostic,
-} from "./diagnostics/deprecation.js";
-export {
-  COMPOSITION_DIAGNOSTIC_CATALOG,
-  duplicateMergeKeyDiagnostic,
-  duplicatePlotLayerDiagnostic,
-  duplicateScaleChannelDiagnostic,
-  isCompositionDiagnostic,
-  isDuplicateMergeKeyDiagnostic,
-  isDuplicatePlotLayerDiagnostic,
-  isDuplicateScaleChannelDiagnostic,
-} from "./diagnostics/composition.js";
-export type {
-  CompositionDiagnostic,
-  CompositionDiagnosticCode,
-  DuplicateMergeKeyDiagnostic,
-  DuplicateMergeKeyKind,
-  DuplicatePlotLayerDiagnostic,
-  DuplicatePlotLayerKind,
-  DuplicateScaleChannelDiagnostic,
-} from "./diagnostics/composition.js";
+export { DEPRECATION_DIAGNOSTIC_CATALOG } from "./diagnostics/deprecation.js";
+export { deprecatedPropDiagnostic } from "./diagnostics/deprecation.js";
+export type { DeprecationDiagnosticCode, PlotDiagnostic } from "./diagnostics/deprecation.js";
+export type { DeprecationDiagnostic } from "./diagnostics/deprecation.js";
+export { isDuplicateScaleChannelDiagnostic } from "./diagnostics/composition.js";
+export { duplicateScaleChannelDiagnostic } from "./diagnostics/composition.js";
+export { COMPOSITION_DIAGNOSTIC_CATALOG } from "./diagnostics/composition.js";
+export { isDuplicatePlotLayerDiagnostic } from "./diagnostics/composition.js";
+export { isDuplicateMergeKeyDiagnostic } from "./diagnostics/composition.js";
+export { duplicatePlotLayerDiagnostic } from "./diagnostics/composition.js";
+export { duplicateMergeKeyDiagnostic, isCompositionDiagnostic } from "./diagnostics/composition.js";
+export type { DuplicateScaleChannelDiagnostic } from "./diagnostics/composition.js";
+export type { DuplicatePlotLayerDiagnostic } from "./diagnostics/composition.js";
+export type { DuplicateMergeKeyDiagnostic } from "./diagnostics/composition.js";
+export type { CompositionDiagnosticCode } from "./diagnostics/composition.js";
+export type { CompositionDiagnostic, DuplicatePlotLayerKind } from "./diagnostics/composition.js";
+export type { DuplicateMergeKeyKind } from "./diagnostics/composition.js";
 
 // Spec surface (builder + canonicalizer + validation)
-export {
-  aes,
-  gg,
-  GGBuilder,
-  coord_equal,
-  coord_fixed,
-  coord_polar,
-  coord_radial,
-  coord_sf,
-  coord_transform,
-  coordEqual,
-  coordFixed,
-  coordPolar,
-  coordRadial,
-  coordSf,
-  coordTransform,
-  isPortable,
-  lintSpec,
-  normalize, // @lifecycle stable-intent
-  dmy,
-  dmy_hm,
-  dmy_hms,
-  dym,
-  dym_hm,
-  dym_hms,
-  fromEpochMilliseconds,
-  fromEpochSeconds,
-  mdy,
-  mdy_hm,
-  mdy_hms,
-  my,
-  myd,
-  myd_hm,
-  myd_hms,
-  parseTemporalFormat,
-  parseTemporalInterval,
-  MAX_TEMPORAL_CANDIDATES,
-  MAX_TEMPORAL_MAJOR_TICKS,
-  MAX_TEMPORAL_MINOR_TICKS,
-  MIN_TEMPORAL_LABEL_GAP_PX,
-  TEMPORAL_INTERVAL_UNITS,
-  TEMPORAL_LABEL_TOKENS,
-  TEMPORAL_WEEKDAYS,
-  temporalIntervalTicks,
-  scaleColorBinned,
-  scaleColorContinuous,
-  scaleColorDate,
-  scaleColorDatetime,
-  scaleColorDiscrete,
-  scaleColorIdentity,
-  scaleColorLog10,
-  scaleColorManual,
-  scaleColorSqrt,
-  scaleColorViridisB,
-  scaleColorViridisC,
-  scaleColorViridisD,
-  scaleContinuousIdentity,
-  scaleDiscreteIdentity,
-  scaleDiscreteManual,
-  scaleType,
-  scale_continuous_identity,
-  scale_discrete_identity,
-  scale_discrete_manual,
-  scale_type,
-  scaleColourBinned,
-  scaleColourContinuous,
-  scaleColourDate,
-  scaleColourDatetime,
-  scaleColourDiscrete,
-  scaleColourIdentity,
-  scaleColourLog10,
-  scaleColourManual,
-  scaleColourSqrt,
-  scaleColourViridisB,
-  scaleColourViridisC,
-  scaleColourViridisD,
-  scaleFillBinned,
-  scaleFillContinuous,
-  scaleFillDate,
-  scaleFillDatetime,
-  scaleFillDiscrete,
-  scaleFillIdentity,
-  scaleFillLog10,
-  scaleFillManual,
-  scaleFillSqrt,
-  scaleFillViridisB,
-  scaleFillViridisC,
-  scaleFillViridisD,
-  scale_color_binned,
-  scale_color_continuous,
-  scale_color_date,
-  scale_color_datetime,
-  scale_color_discrete,
-  scale_color_identity,
-  scale_color_log10,
-  scale_color_manual,
-  scale_color_sqrt,
-  scale_color_viridis_b,
-  scale_color_viridis_c,
-  scale_color_viridis_d,
-  scale_colour_binned,
-  scale_colour_continuous,
-  scale_colour_date,
-  scale_colour_datetime,
-  scale_colour_discrete,
-  scale_colour_identity,
-  scale_colour_log10,
-  scale_colour_manual,
-  scale_colour_sqrt,
-  scale_colour_viridis_b,
-  scale_colour_viridis_c,
-  scale_colour_viridis_d,
-  scale_fill_binned,
-  scale_fill_continuous,
-  scale_fill_date,
-  scale_fill_datetime,
-  scale_fill_discrete,
-  scale_fill_identity,
-  scale_fill_log10,
-  scale_fill_manual,
-  scale_fill_sqrt,
-  scale_fill_viridis_b,
-  scale_fill_viridis_c,
-  scale_fill_viridis_d,
-  scaleXBinned,
-  scaleXContinuous,
-  scaleXDate,
-  scaleXDatetime,
-  scaleXTime,
-  scaleXMonthDay,
-  scaleXDiscrete,
-  scaleXLog10,
-  scaleXReverse,
-  scaleXSqrt,
-  scaleYBinned,
-  scaleYContinuous,
-  scaleYDate,
-  scaleYDatetime,
-  scaleYTime,
-  scaleYMonthDay,
-  scaleYDiscrete,
-  scaleYLog10,
-  scaleYReverse,
-  scaleYSqrt,
-  scale_x_binned,
-  scale_x_continuous,
-  scale_x_date,
-  scale_x_datetime,
-  scale_x_time,
-  scale_x_month_day,
-  scale_x_discrete,
-  scale_x_log10,
-  scale_x_reverse,
-  scale_x_sqrt,
-  scale_y_binned,
-  scale_y_continuous,
-  scale_y_date,
-  scale_y_datetime,
-  scale_y_time,
-  scale_y_month_day,
-  scale_y_discrete,
-  scale_y_log10,
-  scale_y_reverse,
-  scale_y_sqrt,
-  SCALE_CAPABILITIES,
-  guideAxis,
-  guideColorbar,
-  guideColorsteps,
-  guideLegend,
-  guideNone,
-  guides,
-  guide_axis,
-  guide_colorbar,
-  guide_colorsteps,
-  guide_legend,
-  guide_none,
-  ydm,
-  ydm_hm,
-  ydm_hms,
-  ym,
-  ymd,
-  ymd_hm,
-  ymd_hms,
-  yq,
-  SpecValidationError, // @lifecycle stable-intent
-  toPortable,
-  toPortableLossy,
-  validate, // @lifecycle stable-intent
-} from "@ggsvelte/spec";
-export type {
-  A11yMode,
-  Aes,
-  AesInput,
-  AuthoringCellValue,
-  AuthoringColumns,
-  AuthoringDataRef,
-  AuthoringRows,
-  BinnedColorScaleOptions,
-  ChannelValue,
-  ColorScaleOptions,
-  ColorScaleSpec,
-  CoordSpec,
-  CoordFixedOptions,
-  CoordPolarOptions,
-  CoordRadialOptions,
-  CoordRadialReverse,
-  CoordRadialTheta,
-  CoordSfOptions,
-  CoordFixedSpec,
-  CoordRadialSpec,
-  CoordSfSpec,
-  CoordTransformAxisOptions,
-  CoordTransformAxisSpec,
-  CoordTransformName,
-  CoordTransformOptions,
-  CoordTransformSpec,
-  DataInput,
-  BoxplotParams,
-  DataProfile,
-  DataRef,
-  DensityParams,
-  DiscreteColorScaleOptions,
-  FacetInput,
-  FacetScales,
-  FacetSpec,
-  ErrorbarParams,
-  AxisGuideOptions,
-  AxisGuideSpec,
-  ColorbarGuideOptions,
-  ColorbarGuideSpec,
-  ColorstepsGuideOptions,
-  ColorstepsGuideSpec,
-  GuidesSpec,
-  GuideSpec,
-  GuideThemeSpec,
-  LegendGuideOptions,
-  LegendGuideSpec,
-  NoneGuideSpec,
-  LayerInput,
-  LayerSpec,
-  LegendSpec,
-  PointPosition,
-  PortableSpec, // @lifecycle stable-intent
-  PositionParams,
-  PositionScaleSpec,
-  ScaleExpansion,
-  ContinuousPositionScaleOptions,
-  IdentityColorScaleOptions,
-  ManualColorScaleOptions,
-  MultiIdentityScaleOptions,
-  MultiManualScaleOptions,
-  MultiScaleAesthetic,
-  MultiScaleChannel,
-  RecommendedScaleType,
-  ScaleTypeAesthetic,
-  SequentialColorScaleOptions,
-  TemporalColorScaleOptions,
-  TransformedColorScaleOptions,
-  TransformedPositionScaleOptions,
-  ViridisOptionName,
-  ViridisScaleOptions,
-  TemporalDecision,
-  TemporalDisambiguation,
-  TemporalInterval,
-  TemporalIntervalSpec,
-  TemporalIntervalUnit,
-  TemporalKind,
-  TemporalParserName,
-  TemporalParserSpec,
-  TemporalPrecision,
-  TemporalScaleOptions,
-  TemporalWeekStart,
-  DiscretePositionScaleOptions,
-  RenderBackend,
-  RuntimeSpec,
-  ScaleCapability,
-  Scales,
-  SmoothParams,
-  SpecAdvisory,
-  SpecError,
-  SpecInput,
-  StackablePosition,
-  ThemeName,
-  ThemeSpec,
-  ValidateResult,
-} from "@ggsvelte/spec";
-export {
-  LINETYPE_NAMES,
-  POINT_SHAPE_NAMES,
-  STYLE_AESTHETIC_GEOMS,
-  THEME_NAMES,
-  scaleAlpha,
-  scaleAlphaBinned,
-  scaleAlphaContinuous,
-  scaleAlphaDate,
-  scaleAlphaDatetime,
-  scaleAlphaDiscrete,
-  scaleAlphaIdentity,
-  scaleAlphaManual,
-  scaleLinewidth,
-  scaleLinewidthBinned,
-  scaleLinewidthContinuous,
-  scaleLinewidthDate,
-  scaleLinewidthDatetime,
-  scaleLinewidthDiscrete,
-  scaleLinewidthIdentity,
-  scaleLinewidthManual,
-  scaleLinetype,
-  scaleLinetypeBinned,
-  scaleLinetypeDiscrete,
-  scaleLinetypeIdentity,
-  scaleLinetypeManual,
-  scaleShape,
-  scaleShapeBinned,
-  scaleShapeDiscrete,
-  scaleShapeIdentity,
-  scaleShapeManual,
-  scaleSize,
-  scaleSizeBinned,
-  scaleSizeContinuous,
-  scaleSizeDate,
-  scaleSizeDatetime,
-  scaleSizeDiscrete,
-  scaleSizeIdentity,
-  scaleSizeManual,
-  scale_alpha_binned,
-  scale_alpha_continuous,
-  scale_alpha_date,
-  scale_alpha_datetime,
-  scale_alpha_discrete,
-  scale_alpha_identity,
-  scale_alpha_manual,
-  scale_linetype,
-  scale_linetype_binned,
-  scale_linetype_discrete,
-  scale_linetype_identity,
-  scale_linetype_manual,
-  scale_linewidth_binned,
-  scale_linewidth_continuous,
-  scale_linewidth_date,
-  scale_linewidth_datetime,
-  scale_linewidth_discrete,
-  scale_linewidth_identity,
-  scale_linewidth_manual,
-  scale_shape,
-  scale_shape_binned,
-  scale_shape_discrete,
-  scale_shape_identity,
-  scale_shape_manual,
-  scale_size_binned,
-  scale_size_continuous,
-  scale_size_date,
-  scale_size_datetime,
-  scale_size_discrete,
-  scale_size_identity,
-  scale_size_manual,
-} from "@ggsvelte/spec";
-export type {
-  AlphaScaleSpec,
-  BinnedFiniteStyleScaleOptions,
-  DiscreteFiniteStyleScaleOptions,
-  DiscreteNumericStyleScaleOptions,
-  FiniteStyleScaleOptions,
-  IdentityFiniteStyleScaleOptions,
-  IdentityNumericStyleScaleOptions,
-  LinetypeScaleSpec,
-  LinetypeName,
-  PointShapeName,
-  ManualFiniteStyleScaleOptions,
-  ManualNumericStyleScaleOptions,
-  NumericStyleScaleOptions,
-  PositiveStyleScaleSpec,
-  SequentialStyleScaleOptions,
-  ShapeScaleSpec,
-  StyleAesthetic,
-  TemporalNumericStyleScaleOptions,
-} from "@ggsvelte/spec";
+export { aes, coordTransform, coord_equal, coord_radial, coord_transform } from "@ggsvelte/spec";
+export { coordEqual, coordFixed, coordRadial, coord_fixed, coord_polar, gg } from "@ggsvelte/spec";
+export { GGBuilder, coordPolar, coordSf, coord_sf, isPortable, lintSpec } from "@ggsvelte/spec";
+/** @lifecycle stable-intent */ export { normalize } from "@ggsvelte/spec";
+export { MIN_TEMPORAL_LABEL_GAP_PX, scaleXBinned, scale_continuous_identity } from "@ggsvelte/spec";
+export { MAX_TEMPORAL_MAJOR_TICKS, MAX_TEMPORAL_MINOR_TICKS, scaleColorDate } from "@ggsvelte/spec";
+export { MAX_TEMPORAL_CANDIDATES, TEMPORAL_INTERVAL_UNITS, fromEpochSeconds } from "@ggsvelte/spec";
+export { scaleColorBinned, scaleContinuousIdentity, scale_discrete_identity } from "@ggsvelte/spec";
+export { TEMPORAL_WEEKDAYS, scale_color_continuous, scale_colour_continuous } from "@ggsvelte/spec";
+export { scaleColorDatetime, scale_colour_viridis_b, scale_colour_viridis_c } from "@ggsvelte/spec";
+export { fromEpochMilliseconds, parseTemporalFormat, scale_colour_viridis_d } from "@ggsvelte/spec";
+export { TEMPORAL_LABEL_TOKENS, parseTemporalInterval, scaleColorContinuous } from "@ggsvelte/spec";
+export { scaleDiscreteIdentity, scale_color_datetime, temporalIntervalTicks } from "@ggsvelte/spec";
+export { scaleColourContinuous, scale_color_discrete, scale_discrete_manual } from "@ggsvelte/spec";
+export { scale_color_identity, scale_color_viridis_b, scale_color_viridis_c } from "@ggsvelte/spec";
+export { scale_color_viridis_d, scale_colour_datetime, scale_fill_viridis_b } from "@ggsvelte/spec";
+export { scale_colour_discrete, scale_colour_identity, scale_fill_viridis_c } from "@ggsvelte/spec";
+export { scaleDiscreteManual, scale_fill_continuous, scale_fill_viridis_d } from "@ggsvelte/spec";
+export { dmy, scaleColourDatetime, scaleColourDiscrete, scaleColourIdentity } from "@ggsvelte/spec";
+export { dym, scaleColourViridisB, scaleColourViridisC, scaleColourViridisD } from "@ggsvelte/spec";
+export { mdy, scaleFillContinuous, scale_colour_binned, scale_colour_manual } from "@ggsvelte/spec";
+export { myd, scale_fill_datetime, scale_fill_discrete, scale_fill_identity } from "@ggsvelte/spec";
+export { dmy_hm, scaleColorDiscrete, scaleColorIdentity, scaleColorViridisB } from "@ggsvelte/spec";
+export { dym_hm, scaleColorViridisC, scaleColorViridisD, scale_color_binned } from "@ggsvelte/spec";
+export { mdy_hm, scale_color_manual, scale_colour_log10, scale_x_continuous } from "@ggsvelte/spec";
+export { SCALE_CAPABILITIES, dmy_hms, scaleColourBinned, scale_y_continuous } from "@ggsvelte/spec";
+export { scaleColourManual, scaleFillDatetime, scaleFillDiscrete, scaleType } from "@ggsvelte/spec";
+export { guideAxis, scaleFillIdentity, scaleFillViridisB, scaleFillViridisC } from "@ggsvelte/spec";
+export { guideNone, scaleFillViridisD, scale_color_log10, scale_colour_date } from "@ggsvelte/spec";
+export { dym_hms, scale_colour_sqrt, scale_fill_binned, scale_fill_manual } from "@ggsvelte/spec";
+export { scaleColorManual, scale_type, scale_x_month_day, scale_y_month_day } from "@ggsvelte/spec";
+export { scaleColourLog10, scaleYBinned, scale_color_date, scale_color_sqrt } from "@ggsvelte/spec";
+export { scaleXContinuous, scaleYContinuous, scale_fill_log10, scale_x_date } from "@ggsvelte/spec";
+export { scale_x_datetime, scale_x_discrete, scale_x_time, scale_y_datetime } from "@ggsvelte/spec";
+export { guide_colorsteps, scaleColorLog10, scaleFillDate, scale_y_discrete } from "@ggsvelte/spec";
+export { scaleColourDate, scaleColourSqrt, scaleFillBinned, scaleFillManual } from "@ggsvelte/spec";
+export { scale_fill_date, scale_fill_sqrt, scale_x_reverse, scale_y_reverse } from "@ggsvelte/spec";
+export { guideColorsteps, scaleColorSqrt, scaleFillLog10, scaleXDatetime } from "@ggsvelte/spec";
+export { my, scaleXDiscrete, scaleXMonthDay, scaleYDatetime, scaleYMonthDay } from "@ggsvelte/spec";
+export { guide_colorbar, scaleYDiscrete, scale_x_binned, scale_y_binned, ym } from "@ggsvelte/spec";
+export { myd_hm, scaleFillSqrt, scaleXReverse, scaleYReverse, scale_x_log10 } from "@ggsvelte/spec";
+export { guideColorbar, mdy_hms, scale_x_sqrt, scale_y_date, scale_y_log10 } from "@ggsvelte/spec";
+export { guide_legend, scaleXLog10, scaleYLog10, scale_y_sqrt, scale_y_time } from "@ggsvelte/spec";
+export { guideLegend, scaleXDate, scaleXSqrt, scaleXTime, scaleYDate, ydm } from "@ggsvelte/spec";
+export { guide_axis, guide_none, myd_hms, scaleYSqrt, scaleYTime, ydm_hms } from "@ggsvelte/spec";
+export { guides, ydm_hm, ymd, ymd_hm, ymd_hms, yq } from "@ggsvelte/spec";
+/** @lifecycle stable-intent */ export { SpecValidationError } from "@ggsvelte/spec";
+export { toPortable, toPortableLossy } from "@ggsvelte/spec";
+/** @lifecycle stable-intent */ export { validate } from "@ggsvelte/spec";
+export type { CoordTransformAxisOptions, DataRef, DiscreteColorScaleOptions } from "@ggsvelte/spec";
+export type { BinnedColorScaleOptions, ChannelValue, CoordTransformAxisSpec } from "@ggsvelte/spec";
+export type { ColorScaleSpec, ColorstepsGuideOptions, CoordTransformOptions } from "@ggsvelte/spec";
+export type { AuthoringCellValue, ColorbarGuideOptions, ColorstepsGuideSpec } from "@ggsvelte/spec";
+export type { CoordRadialOptions, CoordRadialReverse, CoordTransformName } from "@ggsvelte/spec";
+export type { ColorScaleOptions, CoordTransformSpec, LegendGuideOptions } from "@ggsvelte/spec";
+export type { Aes, ColorbarGuideSpec, CoordFixedOptions, CoordPolarOptions } from "@ggsvelte/spec";
+export type { AuthoringColumns, AuthoringDataRef, CoordRadialTheta } from "@ggsvelte/spec";
+export type { AxisGuideOptions, CoordRadialSpec, CoordSpec, LegendGuideSpec } from "@ggsvelte/spec";
+export type { AuthoringRows, CoordFixedSpec, CoordSfOptions, ErrorbarParams } from "@ggsvelte/spec";
+export type { AxisGuideSpec, BoxplotParams, DensityParams, GuideThemeSpec } from "@ggsvelte/spec";
+export type { CoordSfSpec, DataProfile, NoneGuideSpec, PointPosition } from "@ggsvelte/spec";
+export type { FacetInput, FacetScales, GuidesSpec, LayerInput, LegendSpec } from "@ggsvelte/spec";
+export type { A11yMode, DataInput, FacetSpec, GuideSpec, LayerSpec } from "@ggsvelte/spec";
+export type { AesInput } from "@ggsvelte/spec";
+/** @lifecycle stable-intent */ export type { PortableSpec } from "@ggsvelte/spec";
+export type { TransformedColorScaleOptions, TransformedPositionScaleOptions } from "@ggsvelte/spec";
+export type { ContinuousPositionScaleOptions, DiscretePositionScaleOptions } from "@ggsvelte/spec";
+export type { IdentityColorScaleOptions, SequentialColorScaleOptions } from "@ggsvelte/spec";
+export type { MultiIdentityScaleOptions, Scales, TemporalColorScaleOptions } from "@ggsvelte/spec";
+export type { ManualColorScaleOptions, MultiManualScaleOptions, RuntimeSpec } from "@ggsvelte/spec";
+export type { RecommendedScaleType, ScaleCapability, TemporalDisambiguation } from "@ggsvelte/spec";
+export type { PositionScaleSpec, TemporalIntervalSpec, TemporalIntervalUnit } from "@ggsvelte/spec";
+export type { MultiScaleAesthetic, ScaleTypeAesthetic, TemporalScaleOptions } from "@ggsvelte/spec";
+export type { TemporalParserName, TemporalParserSpec, ViridisScaleOptions } from "@ggsvelte/spec";
+export type { MultiScaleChannel, TemporalPrecision, ViridisOptionName } from "@ggsvelte/spec";
+export type { StackablePosition, TemporalDecision, TemporalWeekStart } from "@ggsvelte/spec";
+export type { PositionParams, ScaleExpansion, SpecError, TemporalInterval } from "@ggsvelte/spec";
+export type { RenderBackend, SmoothParams, TemporalKind, ValidateResult } from "@ggsvelte/spec";
+export type { SpecAdvisory, SpecInput, ThemeName, ThemeSpec } from "@ggsvelte/spec";
+export { THEME_NAMES, scaleLinewidthContinuous, scale_linewidth_continuous } from "@ggsvelte/spec";
+export { LINETYPE_NAMES, scale_linewidth_datetime, scale_linewidth_discrete } from "@ggsvelte/spec";
+export { scaleSizeBinned, scale_linetype_discrete, scale_linewidth_identity } from "@ggsvelte/spec";
+export { POINT_SHAPE_NAMES, scaleLinewidthDatetime, scale_linetype_identity } from "@ggsvelte/spec";
+export { scaleAlphaDatetime, scaleLinewidthDiscrete, scaleLinewidthIdentity } from "@ggsvelte/spec";
+export { scaleAlphaDiscrete, scale_alpha_continuous, scale_linewidth_binned } from "@ggsvelte/spec";
+export { STYLE_AESTHETIC_GEOMS, scaleLinetypeBinned, scale_linewidth_manual } from "@ggsvelte/spec";
+export { scaleAlphaContinuous, scaleLinetypeDiscrete, scaleLinetypeIdentity } from "@ggsvelte/spec";
+export { scaleLinewidthBinned, scale_linetype_binned, scale_linetype_manual } from "@ggsvelte/spec";
+export { scaleLinewidthManual, scale_alpha_datetime, scale_size_continuous } from "@ggsvelte/spec";
+export { scale_alpha_discrete, scale_alpha_identity, scale_linewidth_date } from "@ggsvelte/spec";
+export { scaleLinetypeManual, scale_shape_discrete, scale_shape_identity } from "@ggsvelte/spec";
+export { scaleSizeContinuous, scale_size_datetime, scale_size_discrete } from "@ggsvelte/spec";
+export { scaleAlphaIdentity, scaleLinewidthDate, scale_size_identity } from "@ggsvelte/spec";
+export { scaleShapeDiscrete, scaleShapeIdentity, scale_alpha_binned } from "@ggsvelte/spec";
+export { scale_alpha_manual, scale_shape_binned, scale_shape_manual } from "@ggsvelte/spec";
+export { scaleSize, scaleSizeDatetime, scaleSizeDiscrete, scaleSizeIdentity } from "@ggsvelte/spec";
+export { scaleAlpha, scaleAlphaBinned, scale_size_binned, scale_size_manual } from "@ggsvelte/spec";
+export { scaleAlphaManual, scaleShapeBinned, scaleShapeManual, scale_shape } from "@ggsvelte/spec";
+export { scaleAlphaDate, scaleSizeManual, scale_alpha_date, scale_size_date } from "@ggsvelte/spec";
+export { scaleLinetype, scaleLinewidth, scaleSizeDate, scale_linetype } from "@ggsvelte/spec";
+export { scaleShape } from "@ggsvelte/spec";
+export type { DiscreteNumericStyleScaleOptions, SequentialStyleScaleOptions } from "@ggsvelte/spec";
+export type { IdentityNumericStyleScaleOptions, NumericStyleScaleOptions } from "@ggsvelte/spec";
+export type { FiniteStyleScaleOptions, TemporalNumericStyleScaleOptions } from "@ggsvelte/spec";
+export type { DiscreteFiniteStyleScaleOptions, PositiveStyleScaleSpec } from "@ggsvelte/spec";
+export type { IdentityFiniteStyleScaleOptions, LinetypeScaleSpec } from "@ggsvelte/spec";
+export type { BinnedFiniteStyleScaleOptions, ManualNumericStyleScaleOptions } from "@ggsvelte/spec";
+export type { AlphaScaleSpec, ManualFiniteStyleScaleOptions, PointShapeName } from "@ggsvelte/spec";
+export type { LinetypeName, ShapeScaleSpec, StyleAesthetic } from "@ggsvelte/spec";
 
 // Core surface (pure renderer + pipeline)
-export {
-  PipelineError,
-  installTemporal, // @lifecycle experimental
-  planStrata,
-  registerAbline, // @lifecycle experimental
-  registerAlign, // @lifecycle experimental
-  registerAll, // @lifecycle experimental
-  registerBasic, // @lifecycle experimental
-  registerBin, // @lifecycle experimental
-  registerBin2d, // @lifecycle experimental
-  registerBoxplot, // @lifecycle experimental
-  registerConnect, // @lifecycle experimental
-  registerContour, // @lifecycle experimental
-  registerCrossbar, // @lifecycle experimental
-  registerCurve, // @lifecycle experimental
-  registerDensity, // @lifecycle experimental
-  registerDensity2d, // @lifecycle experimental
-  registerDensity2dFilled, // @lifecycle experimental
-  registerDotplot, // @lifecycle experimental
-  registerEcdf, // @lifecycle experimental
-  registerEllipse, // @lifecycle experimental
-  registerErrorbar, // @lifecycle experimental
-  registerFunction, // @lifecycle experimental
-  registerHex, // @lifecycle experimental
-  registerLinerange, // @lifecycle experimental
-  registerManual, // @lifecycle experimental
-  registerMap, // @lifecycle experimental
-  registerPointrange, // @lifecycle experimental
-  registerPolygon, // @lifecycle experimental
-  registerQq, // @lifecycle experimental
-  registerQqLine, // @lifecycle experimental
-  registerQuantile, // @lifecycle experimental
-  registerRaster, // @lifecycle experimental
-  registerRug, // @lifecycle experimental
-  registerSf, // @lifecycle experimental
-  registerSfLabel, // @lifecycle experimental
-  registerSfText, // @lifecycle experimental
-  registerSmooth, // @lifecycle experimental
-  registerSpoke, // @lifecycle experimental
-  registerSummary, // @lifecycle experimental
-  registerSummaryBin, // @lifecycle experimental
-  registerSummaryRolling, // @lifecycle experimental
-  registerTile, // @lifecycle experimental
-  registerUnique, // @lifecycle experimental
-  registerViolin, // @lifecycle experimental
-  renderToSVGString, // @lifecycle stable-intent
-  runPipeline,
-} from "@ggsvelte/core";
-export type {
-  Advisory,
-  AxisGuidePlan,
-  AxisGuideTick,
-  ColorbarGuidePlan,
-  ColorbarGuideTick,
-  ColorstepsGuidePlan,
-  ColorstepsGuideStep,
-  DiscreteGuideEntry,
-  DiscreteGuidePlan,
-  GuidePlan,
-  LayerBackend,
-  MappedField,
-  PipelineWarning,
-  RenderModel,
-  RenderSVGOptions,
-  ResolvedColorScale,
-  ResolvedStyleScale,
-  PointShape,
-  Linetype,
-  RunOptions,
-  ScaleDecision,
-  ScaleDiagnostic,
-  ScaleDiagnosticFix,
-  ScaleState,
-  Scene,
-  Stratum,
-  ThemeTokens,
-} from "@ggsvelte/core";
+export { registerDensity2d, registerDensity2dFilled, registerSummaryRolling } from "@ggsvelte/core";
+export { registerLinerange, registerPointrange, registerSummaryBin } from "@ggsvelte/core";
+export { registerCrossbar, registerEcdf, registerErrorbar, registerFunction } from "@ggsvelte/core";
+export { installTemporal, registerAbline, registerBoxplot, registerQuantile } from "@ggsvelte/core";
+export { registerConnect, registerContour, registerDensity, registerDotplot } from "@ggsvelte/core";
+export { registerEllipse, registerPolygon, registerSfLabel, registerSummary } from "@ggsvelte/core";
+export { registerManual, registerQqLine, registerRaster, registerSfText } from "@ggsvelte/core";
+export { PipelineError, registerSmooth, registerUnique, registerViolin } from "@ggsvelte/core";
+export { registerAlign, registerBasic, registerBin2d, registerCurve } from "@ggsvelte/core";
+export { registerAll, registerBin, registerHex, registerSpoke, registerTile } from "@ggsvelte/core";
+export { planStrata, registerMap, registerQq, registerRug, registerSf } from "@ggsvelte/core";
+/** @lifecycle stable-intent */ export { renderToSVGString } from "@ggsvelte/core";
+export { runPipeline } from "@ggsvelte/core";
+export type { ColorstepsGuidePlan, ColorstepsGuideStep, DiscreteGuideEntry } from "@ggsvelte/core";
+export type { ResolvedColorScale, ResolvedStyleScale, ScaleDiagnosticFix } from "@ggsvelte/core";
+export type { ColorbarGuidePlan, ColorbarGuideTick, DiscreteGuidePlan } from "@ggsvelte/core";
+export type { GuidePlan, PipelineWarning, RenderSVGOptions, ScaleDiagnostic } from "@ggsvelte/core";
+export type { AxisGuidePlan, AxisGuideTick, LayerBackend, ScaleDecision } from "@ggsvelte/core";
+export type { MappedField, PointShape, RenderModel, RunOptions, ThemeTokens } from "@ggsvelte/core";
+export type { Advisory, Linetype, ScaleState, Scene, Stratum } from "@ggsvelte/core";
