+++ scripts/gen-geom-children.test.ts
@@ -16,6 +16,7 @@ import {
   SHELL_MANIFEST,
   SHELL_REGISTRATION,
 } from "./gen-geom-children-manifest.ts";
+import { extractExports } from "./gen-lifecycle.ts";
 import {
   GENERATED_HEADER,
   INDEX_PATH,
@@ -156,12 +157,33 @@ describe("index region rewrite", () => {
     expect(out).toBe(`// keep\n${REGION_START}\nnew\n${REGION_END}\n// after\n`);
   });
 
-  it("emits one lifecycle-tagged export per shell (49)", () => {
+  it("emits one compact lifecycle-tagged export line per shell (49)", () => {
     const region = renderIndexRegion();
     expect(region.startsWith(REGION_START)).toBe(true);
     expect(region.endsWith(REGION_END)).toBe(true);
-    const exportCount = (region.match(/^export \{ default as /gm) ?? []).length;
-    expect(exportCount).toBe(49);
+    const body = region.split("\n").slice(1, -1);
+    expect(body).toHaveLength(49);
+    const compact =
+      /^\/\*\* @lifecycle stable-intent \*\/ export \{ default as (\w+) \} from "\.\/geoms\/\w+\.svelte";$/;
+    const names = body.map((line) => {
+      const m = compact.exec(line);
+      expect(m, `non-compact export line: ${line}`).not.toBeNull();
+      return m![1]!;
+    });
+    expect(new Set(names).size).toBe(49);
+    expect(names).toEqual(SHELL_MANIFEST.map((s) => s.component));
+  });
+
+  it("gen-lifecycle parses the compact region as stable-intent value exports", () => {
+    const exports = extractExports(
+      `// @lifecycle-default experimental\n${renderIndexRegion()}\n`,
+      INDEX_PATH,
+    );
+    expect(exports).toHaveLength(49);
+    for (const e of exports) {
+      expect(e.kind).toBe("value");
+      expect(e.lifecycle).toBe("stable-intent");
+    }
   });
 });
