+++ scripts/gen-scale-children.ts
@@ -88,12 +88,14 @@ export function renderShell(spec: ShellSpec): string {
   ].join("\n");
 }
 
-/** One lifecycle-tagged export line (form gen-lifecycle already parses). */
+/**
+ * One compact lifecycle-tagged export line (form gen-lifecycle already
+ * parses): an inline `@lifecycle` JSDoc marker, then the single-name
+ * statement on the same line. oxfmt hugs single-specifier exports, so the
+ * line is stable at any length.
+ */
 function exportLine(asName: string, fileComponent: string): string {
-  return [
-    `/** @lifecycle stable-intent */`,
-    `export { default as ${asName} } from "./scale/${fileComponent}.svelte";`,
-  ].join("\n");
+  return `/** @lifecycle stable-intent */ export { default as ${asName} } from "./scale/${fileComponent}.svelte";`;
 }
 
 /** Render the full delimited index region body (markers included). Pure. */
