+++ scripts/guards/max-loc-exclusions.txt
@@ -10,7 +10,6 @@ packages/svelte/src/lib/data/mpg.ts 3099 # pure data corpus, no logic — perman
 examples/interaction/tooltip/data.ts 3019 # pure data corpus, no logic — permanently exempt
 scripts/ci-routing.test.ts 1571 # legacy oversize; pending refactor slice S7
 tests/evals/model.ts 1266 # legacy oversize; pending refactor slice S7
-packages/svelte/src/lib/index.ts 1086 # legacy oversize; pending refactor slice S5
 examples/hex/basic/data.ts 1014 # pure data corpus, no logic — permanently exempt
 scripts/release-wiring.test.ts 1005 # legacy oversize; pending refactor slice S7
 examples/density/overlay/data.ts 948 # pure data corpus, no logic — permanently exempt
@@ -18,13 +17,11 @@ examples/errorbar/summary-bin/data.ts 945 # pure data corpus, no logic — perma
 examples/point/count/data.ts 943 # pure data corpus, no logic — permanently exempt
 packages/svelte/tests/inspection/resolve-snapshot.test.ts 940 # legacy oversize; pending refactor slice S6
 packages/svelte/src/lib/data/kyoto-sakura.ts 891 # pure data corpus, no logic — permanently exempt
-packages/svelte/src/lib/plot-engine.svelte.ts 883 # legacy oversize; pending refactor slice S5
 scripts/sakura-lesson.test.ts 870 # legacy oversize; pending refactor slice S7
 apps/docs/src/lib/theme-specimens/data.ts 821 # pure data corpus, no logic — permanently exempt
 examples/point/stat-unique/data.ts 796 # pure data corpus, no logic — permanently exempt
 examples/line/ecdf/data.ts 793 # pure data corpus, no logic — permanently exempt
 packages/svelte/tests/inspection/coordinator.test.ts 736 # legacy oversize; pending refactor slice S6
-packages/svelte/src/lib/surface/surface-state.svelte.ts 717 # legacy oversize; pending refactor slice S5
 scripts/quickstart/steps.ts 698 # legacy oversize; pending refactor slice S7
 examples/path/trajectory/data.ts 697 # pure data corpus, no logic — permanently exempt
 packages/svelte/tests/inspection/display-members.test.ts 695 # legacy oversize; pending refactor slice S6
@@ -35,18 +32,13 @@ scripts/ci-routing/routing.ts 638 # legacy oversize; pending refactor slice S7
 packages/svelte/tests/interval/consumption.test.ts 628 # legacy oversize; pending refactor slice S6
 packages/svelte/tests/legend/filter-state.test.ts 617 # legacy oversize; pending refactor slice S6
 examples/density/kde-2d/data.ts 611 # pure data corpus, no logic — permanently exempt
-packages/svelte/src/lib/inspection/inspection-state.svelte.ts 585 # legacy oversize; pending refactor slice S5
-packages/svelte/src/lib/interval/interval-state.svelte.ts 583 # legacy oversize; pending refactor slice S5
 scripts/gen-benchmark-charts.ts 583 # legacy oversize; pending refactor slice S7
 packages/svelte/tests/legend/focus-entry-key-index-pure.test.ts 581 # legacy oversize; pending refactor slice S6
 packages/svelte/tests/legend/focus-interaction.test.ts 577 # legacy oversize; pending refactor slice S6
-packages/svelte/src/lib/GGPlot.svelte 570 # legacy oversize; pending refactor slice S5
 scripts/skill-content.test.ts 563 # legacy oversize; pending refactor slice S7
 packages/svelte/tests/inspection/inspection-state.pin-queue.test.ts 544 # legacy oversize; pending refactor slice S6
-packages/svelte/src/lib/legend/focus-state.svelte.ts 543 # legacy oversize; pending refactor slice S5
 packages/svelte/tests/runtime/semantic-candidate-projection.test.ts 543 # legacy oversize; pending refactor slice S6
 examples/point/gradient-continuous/data.ts 530 # pure data corpus, no logic — permanently exempt
-packages/svelte/src/lib/scene/Batch.svelte 526 # legacy oversize; pending refactor slice S5
 benchmarks/competitive/fixtures/main.ts 523 # legacy oversize; pending refactor slice S7
 packages/svelte/tests/selection/selection.test.ts 521 # legacy oversize; pending refactor slice S6
 scripts/gen-llms.test.ts 521 # legacy oversize; pending refactor slice S7
